httptransport: Refactor Matcher to align with indexer
These changes create a dedicated matcher handler that
is responsible for the all endpoints the matcher services.
