+++ local-dev/grafana/provisioning/dashboards/dashboard.json
@@ -21,7 +21,7 @@
   "editable": true,
   "gnetId": null,
   "graphTooltip": 1,
-  "iteration": 1645816279286,
+  "iteration": 1646408105440,
   "links": [],
   "panels": [
     {
@@ -2082,7 +2082,7 @@
       "targets": [
         {
           "exemplar": true,
-          "expr": "sum by (code) (rate(clair_http_matcher_api_v1_vulnerability_report_request_total[$rate]))",
+          "expr": "sum by (code) (rate(clair_http_matcherv1_request_total{handler=\"/matcher/api/v1/vulnerability_report/\"}[$rate]))",
           "instant": false,
           "interval": "",
           "legendFormat": "Status {{ code }} ",
@@ -2125,7 +2125,7 @@
       "targets": [
         {
           "exemplar": true,
-          "expr": "sum(rate(clair_http_matcher_api_v1_vulnerability_report_request_duration_seconds_bucket [5m])) by (le)",
+          "expr": "sum(rate(clair_http_matcherv1_request_duration_seconds_bucket{handler=\"/matcher/api/v1/vulnerability_report/\"}[5m])) by (le)",
           "format": "heatmap",
           "instant": false,
           "interval": "",
@@ -3303,7 +3303,7 @@
     ]
   },
   "time": {
-    "from": "now-6h",
+    "from": "now-1h",
     "to": "now"
   },
   "timepicker": {},
